requirements.txt
a2wsgi==1.10.4
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.6
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
starlette==0.37.2
swagger_ui_bundle==1.1.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.1
